feat: carousel configuration refactored and nestjs skill added
